others/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.13.0
altair==5.1.2
ansiwrap==0.8.4
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.6
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
augly==1.0.0
autopage==0.5.2
autopep8==2.0.2
av==11.0.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.0
black==23.7.0
bleach==6.0.0
blessed==1.20.0
cachetools==5.3.1
certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cliff==4.4.0
clip==0.2.0
cloudpickle==2.2.1
cmaes==0.10.0
cmake==3.26.0
cmd2==2.4.3
colorama==0.4.6
colorlog==6.8.0
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.0
dask==2023.7.1
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
deepspeed==0.12.3
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.19.3
dill==0.3.7
distributed==2023.7.1
docker-pycreds==0.4.0
docstring-parser==0.15
docstring-to-markdown==0.12
einops==0.6.1
einops-exts==0.0.4
elyra-examples-kfp-catalog==0.1.0
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
face-detection==0.2.2
faiss-cpu==1.7.4
fastapi==0.104.1
fastjsonschema==2.18.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.10.0
fire==0.5.0
flake8==6.0.0
Flask==2.2.3
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.3.3
fschat==0.2.32
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==1.12.11
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gpustat==1.1
gradio==3.35.2
gradio_client==0.2.9
greenlet==2.0.2
grpcio==1.60.0
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.19.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
inflect==7.0.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.25.0
ipymonaco==0.0.23a0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonlines==4.0.0
jsonpath-ng==1.5.3
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-resource-usage==1.0.0
jupyter-server-mathjax==0.2.6
jupyter-ydoc==0.2.5
jupyter_ai==2.0.0
jupyter_ai_magics==2.0.0
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_packaging==0.12.3
jupyter_server==2.7.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.5
jupyterlab-lsp==4.2.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kfp==1.8.22
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kfp-tekton==1.7.1
kfp-tekton-server-api==1.5.0
kiwisolver==1.4.4
kubernetes==25.3.0
langchain==0.0.220
langchainplus-sdk==0.0.20
lazy-object-proxy==1.9.0
lazy_loader==0.3
lightning==2.1.2
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lit==15.0.7
locket==1.0.0
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
minio==7.1.15
mistune==3.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbdime==3.2.1
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.0
nh3==0.2.14
ninja==1.11.1.1
nltk==3.8.1
notebook==6.5.5
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.2
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.74
optuna==2.10.1
orjson==3.9.10
overrides==7.3.1
packaging==23.0
pandas==2.1.3
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.2
pbr==6.0.0
peft==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
pluggy==1.2.0
ply==3.11
portalocker==2.7.0
posthog==3.0.2
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.12
pydocstyle==6.3.0
pydub==0.25.1
pyflakes==3.0.1
PyGithub==1.59.1
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.5
PyNaCl==1.5.0
pynvml==11.5.0
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-json-logger==2.0.7
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
python-magic==0.4.27
python-multipart==0.0.6
pytoolconfig==1.2.5
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
ray==2.8.0
referencing==0.30.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rootutils==1.0.7
rope==1.9.0
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.39.1
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
stevedore==5.1.0
strip-hints==0.1.10
svgwrite==1.4.3
sympy==1.11.1
tabulate==0.9.0
tblib==2.0.0
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
textwrap3==0.9.2
threadpoolctl==3.2.0
tifffile==2023.8.12
tiktoken==0.4.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.4.3
torchaudio==2.0.1
torchmetrics==1.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.36.0
transformers-stream-generator==0.0.4
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
ujson==5.8.0
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.15
uvicorn==0.24.0.post1
uvloop==0.19.0
vllm==0.2.1.post1
wandb==0.16.1
watchdog==3.0.0
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
Werkzeug==2.2.3
whatthepatch==1.0.5
widgetsnbextension==4.0.8
wrapt==1.15.0
xformers==0.0.22
y-py==0.6.0
yapf==0.40.1
yarl==1.8.2
yaspin==2.5.0
ypy-websocket==0.8.4
zict==3.0.0
zipp==3.16.2
